- Update USR's code 2019 (thank's Andrew)
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@6625 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 3361a902b122e817704f3d9f1c78bf8ab3e66b38
Former-commit-id: 868cc91bb9677e1a14395656cd2b0c39bf6f084c

+++ src/drivers/usr/src/linalg.h
@@ -20,7 +20,7 @@
 #ifndef _LINALG_H_
 #define _LINALG_H_
 
-inline double sign(double d)
+inline float sign(float d)
 {
     return (d >= 0.0) ? 1.0f : -1.0f;
 }
@@ -32,29 +32,28 @@ public:
     /* constructors */
     v2d() {}
     v2d(const v2d &src) { this->x = src.x; this->y = src.y; }
-    v2d(double x, double y) { this->x = x; this->y = y; }
+    v2d(float x, float y) { this->x = x; this->y = y; }
 
     /* operators */
-    v2d& operator=(const v2d &src);           /* assignment */
-    v2d operator+(const v2d &src) const;      /* addition */
-    v2d operator-(void) const;                /* negation */
-    v2d operator-(const v2d &src) const;      /* subtraction */
-    v2d operator*(const double s) const;      /* multiply with scalar */
-    double operator*(const v2d &src) const;   /* dot product */
-    friend v2d operator*(const double s, const v2d & src);
+    v2d& operator=(const v2d &src);         /* assignment */
+    v2d operator+(const v2d &src) const;    /* addition */
+    v2d operator-(void) const;              /* negation */
+    v2d operator-(const v2d &src) const;    /* subtraction */
+    v2d operator*(const float s) const;     /* multiply with scalar */
+    float operator*(const v2d &src) const;  /* dot product */
+    friend v2d operator*(const float s, const v2d & src);
 
     /* methods */
-    double len(void) const;
+    float len(void) const;
     void normalize(void);
-    double dist(const v2d &p) const;
-    double cosalpha(const v2d &p2, const v2d &center) const;
-    v2d rotate(const v2d &c, double arc) const;
+    float dist(const v2d &p) const;
+    float cosalpha(const v2d &p2, const v2d &center) const;
+    v2d rotate(const v2d &c, float arc) const;
 
     /* data */
-    double x;
-    double y;
+    float x;
+    float y;
 };
-
 
 /* assignment */
 inline v2d& v2d::operator=(const v2d &src)
@@ -62,13 +61,11 @@ inline v2d& v2d::operator=(const v2d &src)
     x = src.x; y = src.y; return *this;
 }
 
-
 /* add *this + src (vector addition) */
 inline v2d v2d::operator+(const v2d &src) const
 {
     return v2d(x + src.x, y + src.y);
 }
-
 
 /* negation of *this */
 inline v2d v2d::operator-(void) const
@@ -76,108 +73,100 @@ inline v2d v2d::operator-(void) const
     return v2d(-x, -y);
 }
 
-
 /* compute *this - src (vector subtraction) */
 inline v2d v2d::operator-(const v2d &src) const
 {
     return v2d(x - src.x, y - src.y);
 }
 
-
 /* scalar product */
-inline double v2d::operator*(const v2d &src) const
+inline float v2d::operator*(const v2d &src) const
 {
     return src.x*x + src.y*y;
 }
 
-
-/* multiply vector with scalar (v2d*double) */
-inline v2d v2d::operator*(const double s) const
+/* multiply vector with scalar (v2d*float) */
+inline v2d v2d::operator*(const float s) const
 {
     return v2d(s*x, s*y);
 }
 
-
-/* multiply scalar with vector (double*v2d) */
-inline v2d operator*(const double s, const v2d & src)
+/* multiply scalar with vector (float*v2d) */
+inline v2d operator*(const float s, const v2d & src)
 {
     return v2d(s*src.x, s*src.y);
 }
 
-
 /* compute cosine of the angle between vectors *this-c and p2-c */
-inline double v2d::cosalpha(const v2d &p2, const v2d &c) const
+inline float v2d::cosalpha(const v2d &p2, const v2d &c) const
 {
     v2d l1 = *this-c;
     v2d l2 = p2 - c;
+
     return (l1*l2)/(l1.len()*l2.len());
 }
 
-
 /* rotate vector arc radians around center c */
-inline v2d v2d::rotate(const v2d &c, double arc) const
+inline v2d v2d::rotate(const v2d &c, float arc) const
 {
     v2d d = *this-c;
-    double sina = sin(arc), cosa = cos(arc);
+    float sina = sin(arc), cosa = cos(arc);
+
     return c + v2d(d.x*cosa-d.y*sina, d.x*sina+d.y*cosa);
 }
 
-
 /* compute the length of the vector */
-inline double v2d::len(void) const
+inline float v2d::len(void) const
 {
     return sqrt(x*x+y*y);
 }
 
-
 /* distance between *this and p */
-inline double v2d::dist(const v2d &p) const
-{ 
+inline float v2d::dist(const v2d &p) const
+{
     return sqrt((p.x-x)*(p.x-x)+(p.y-y)*(p.y-y));
 }
-
 
 /* normalize the vector */
 inline void v2d::normalize(void)
 {
-    double l = this->len();
+    float l = this->len();
     x /= l; y /= l;
 }
-
 
 class Straight
 {
 public:
     /* constructors */
     Straight() {}
-    Straight(double x, double y, double dx, double dy)
+    Straight(float x, float y, float dx, float dy)
     {  p.x = x; p.y = y; d.x = dx; d.y = dy; d.normalize(); }
     Straight(const v2d &anchor, const v2d &dir)
     { p = anchor; d = dir; d.normalize(); }
 
     /* methods */
     v2d intersect(const Straight &s) const;
-    double dist(const v2d &p) const;
+    float dist(const v2d &p) const;
 
     /* data */
     v2d p;          /* point on the straight */
     v2d d;          /* direction of the straight */
 };
 
-
 /* intersection point of *this and s */
 inline v2d Straight::intersect(const Straight &s) const
 {
-    double t = -(d.x*(s.p.y-p.y)+d.y*(p.x-s.p.x))/(d.x*s.d.y-d.y*s.d.x);
+    float t = -(d.x*(s.p.y-p.y)+d.y*(p.x-s.p.x))/(d.x*s.d.y-d.y*s.d.x);
+
     return s.p + s.d*t;
 }
 
-
 /* distance of point s from straight *this */
-inline double Straight::dist(const v2d &s) const
+inline float Straight::dist(const v2d &s) const
 {
     v2d d1 = s - p;
     v2d d3 = d1 - d*d1*d;
+
     return d3.len();
 }
 
@@ -208,67 +197,56 @@ public:
     void dirVector(const v3d* b, v3d* r);     /* r := this - b */
 };
 
-
 inline v3d & v3d::operator=(const v3d &src)
 {
     x = src.x; y = src.y; z = src.z; return *this;
 }
-
 
 inline v3d v3d::operator-(void)
 {
     return v3d(-this->x, -this->y, -this->z);
 }
 
-
 inline v3d v3d::operator-(const v3d &src)
 {
     return v3d(this->x - src.x, this->y - src.y, this->z - src.z);
 }
-
 
 inline v3d v3d::operator+(const v3d &src)
 {
     return v3d(this->x + src.x, this->y + src.y, this->z + src.z);
 }
 
-
 inline v3d v3d::operator*(const double s)
 {
     return v3d(s*this->x, s*this->y, s*this->z);
 }
-
 
 inline double v3d::operator*(const v3d &src)
 {
     return this->x*src.x + this->y*src.y + this->z*src.z;
 }
 
-
 inline v3d operator*(const double s, const v3d & src)
 {
     return v3d(s*src.x, s*src.y, s*src.z);
 }
-
 
 inline v3d v3d::operator/(const double s)
 {
     return v3d(this->x/s, this->y/s, this->z/s);
 }
 
-
 inline double v3d::len(void)
 {
     return sqrt(x*x + y*y + z*z);
 }
-
 
 inline void v3d::normalize(void)
 {
     double len = this->len();
     x /= len; y /= len; z /= len;
 }
-
 
 inline void v3d::crossProduct(const v3d* b, v3d* r)
 {
@@ -282,4 +260,4 @@ inline void v3d::dirVector(const v3d* b, v3d* r)
     r->x = this->x - b->x; r->y = this->y - b->y; r->z = this->z - b->z;
 }
 
-#endif // _LINALG_H_
+#endif
